Update x-ago timestamps on page load and every 5 minutes.
