+++ app/lib/shared/utils.dart
@@ -38,28 +38,6 @@ final DateFormat shortDateFormat = DateFormat.yMMMd();
 final jsonUtf8Encoder = JsonUtf8Encoder();
 final utf8JsonDecoder = utf8.decoder.fuse(json.decoder);
 
-/// Formats an [age] duration with `<amount> <unit> ago` or `in the last hour`.
-String formatXAgo(Duration age) {
-  if (age.inDays > 365 * 2) {
-    final years = age.inDays ~/ 365;
-    return '$years years ago';
-  }
-  if (age.inDays > 30 * 2) {
-    final months = age.inDays ~/ 30;
-    return '$months months ago';
-  }
-  if (age.inDays > 1) {
-    return '${age.inDays} days ago';
-  }
-  if (age.inHours > 1) {
-    return '${age.inHours} hours ago';
-  }
-  if (age.inHours == 1) {
-    return '${age.inHours} hour ago';
-  }
-  return 'in the last hour';
-}
-
 Future<T> withTempDirectory<T>(Future<T> Function(Directory dir) func,
     {String prefix = 'dart-tempdir'}) {
   return Directory.systemTemp.createTemp(prefix).then((Directory dir) {
