+++ app/test/frontend/golden/my_liked_packages.html
@@ -148,7 +148,7 @@
                   <p>user@pub.dev</p>
                   <p>
                     Joined
-                    <a class="-x-ago" href="" title="on %%user-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                    <a class="-x-ago" href="" title="on %%user-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                   </p>
                 </div>
               </div>
@@ -193,7 +193,7 @@
                       </div>
                       <p class="packages-metadata">
                         Liked
-                        <a class="-x-ago" href="" title="on %%liked1-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                        <a class="-x-ago" href="" title="on %%liked1-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                       </p>
                     </div>
                     <div class="packages-item">
@@ -211,7 +211,7 @@
                       </div>
                       <p class="packages-metadata">
                         Liked
-                        <a class="-x-ago" href="" title="on %%liked1-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                        <a class="-x-ago" href="" title="on %%liked1-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                       </p>
                     </div>
                   </div>
