+++ app/test/frontend/golden/my_packages.html
@@ -147,7 +147,7 @@
                   <p>user@pub.dev</p>
                   <p>
                     Joined
-                    <a class="-x-ago" href="" title="on %%oxygen-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                    <a class="-x-ago" href="" title="on %%oxygen-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                   </p>
                 </div>
               </div>
@@ -226,11 +226,11 @@
                               v
                               <a href="/packages/oxygen">1.2.0</a>
                               (
-                              <a class="-x-ago" href="" title="%%oxygen-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                              <a class="-x-ago" href="" title="%%oxygen-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                               ) /
                               <a href="/packages/oxygen/versions/2.0.0-dev">2.0.0-dev</a>
                               (
-                              <a class="-x-ago" href="" title="%%oxygen-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                              <a class="-x-ago" href="" title="%%oxygen-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                               )
                             </span>
                             <span class="packages-metadata-block">
@@ -302,7 +302,7 @@
                               v
                               <a href="/packages/neon">1.0.0</a>
                               (
-                              <a class="-x-ago" href="" title="%%oxygen-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                              <a class="-x-ago" href="" title="%%oxygen-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                               )
                             </span>
                             <span class="packages-metadata-block">
