+++ app/test/frontend/golden/my_publishers.html
@@ -147,7 +147,7 @@
                   <p>user@pub.dev</p>
                   <p>
                     Joined
-                    <a class="-x-ago" href="" title="on %%user-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                    <a class="-x-ago" href="" title="on %%user-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                   </p>
                 </div>
               </div>
@@ -182,7 +182,7 @@
                       </h3>
                       <p>
                         Registered
-                        <a class="-x-ago" href="" title="on %%publisher-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                        <a class="-x-ago" href="" title="on %%publisher-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                         .
                       </p>
                     </div>
