+++ app/test/frontend/golden/pkg_activity_log_page.html
@@ -154,7 +154,7 @@
                 <div class="metadata">
                   Published
                   <span>
-                    <a class="-x-ago" href="" title="%%published-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                    <a class="-x-ago" href="" title="%%published-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                   </span>
                   • Latest:
                   <span>
@@ -257,7 +257,7 @@
                     <tbody>
                       <tr>
                         <td class="date">
-                          <a class="-x-ago" href="" title="%%published-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                          <a class="-x-ago" href="" title="%%published-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                         </td>
                         <td class="summary">
                           <div class="markdown-body">
@@ -271,7 +271,7 @@
                       </tr>
                       <tr>
                         <td class="date">
-                          <a class="-x-ago" href="" title="%%published-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                          <a class="-x-ago" href="" title="%%published-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                         </td>
                         <td class="summary">
                           <div class="markdown-body">
@@ -289,7 +289,7 @@
                       </tr>
                       <tr>
                         <td class="date">
-                          <a class="-x-ago" href="" title="%%published-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                          <a class="-x-ago" href="" title="%%published-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                         </td>
                         <td class="summary">
                           <div class="markdown-body">
@@ -307,7 +307,7 @@
                       </tr>
                       <tr>
                         <td class="date">
-                          <a class="-x-ago" href="" title="%%published-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                          <a class="-x-ago" href="" title="%%published-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                         </td>
                         <td class="summary">
                           <div class="markdown-body">
@@ -325,7 +325,7 @@
                       </tr>
                       <tr>
                         <td class="date">
-                          <a class="-x-ago" href="" title="%%activity-4-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                          <a class="-x-ago" href="" title="%%activity-4-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                         </td>
                         <td class="summary">
                           <div class="markdown-body">
