+++ app/test/frontend/golden/pkg_index_page.html
@@ -476,11 +476,11 @@
                       v
                       <a href="/packages/oxygen">1.2.0</a>
                       (
-                      <a class="-x-ago" href="" title="%%oxygen-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                      <a class="-x-ago" href="" title="%%oxygen-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                       ) /
                       <a href="/packages/oxygen/versions/2.0.0-dev">2.0.0-dev</a>
                       (
-                      <a class="-x-ago" href="" title="%%oxygen-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                      <a class="-x-ago" href="" title="%%oxygen-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                       )
                     </span>
                     <span class="packages-metadata-block">
@@ -552,7 +552,7 @@
                       v
                       <a href="/packages/flutter_titanium">1.10.0</a>
                       (
-                      <a class="-x-ago" href="" title="%%oxygen-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                      <a class="-x-ago" href="" title="%%oxygen-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                       )
                     </span>
                     <span class="packages-metadata-block">
