+++ app/test/frontend/golden/pkg_show_page_retracted_non_retracted_version.html
@@ -128,7 +128,7 @@
                 <div class="metadata">
                   Published
                   <span>
-                    <a class="-x-ago" href="" title="%%published-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                    <a class="-x-ago" href="" title="%%published-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                   </span>
                 </div>
                 <div class="detail-tags-and-like">
