+++ app/test/frontend/golden/publisher_list_page.html
@@ -123,7 +123,7 @@
           </h3>
           <p>
             Registered
-            <a class="-x-ago" href="" title="on %%example-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+            <a class="-x-ago" href="" title="on %%example-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
             .
           </p>
         </div>
@@ -133,7 +133,7 @@
           </h3>
           <p>
             Registered
-            <a class="-x-ago" href="" title="on %%other-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+            <a class="-x-ago" href="" title="on %%other-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
             .
           </p>
         </div>
