+++ app/test/frontend/golden/publisher_packages_page.html
@@ -131,7 +131,7 @@
                   </p>
                   <p>
                     Publisher registered
-                    <a class="-x-ago" href="" title="on %%neon-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                    <a class="-x-ago" href="" title="on %%neon-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                   </p>
                 </div>
               </div>
@@ -228,7 +228,7 @@
                               v
                               <a href="/packages/neon">1.0.0</a>
                               (
-                              <a class="-x-ago" href="" title="%%neon-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                              <a class="-x-ago" href="" title="%%neon-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                               )
                             </span>
                             <span class="packages-metadata-block">
@@ -302,7 +302,7 @@
                               v
                               <a href="/packages/flutter_titanium">1.10.0</a>
                               (
-                              <a class="-x-ago" href="" title="%%neon-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                              <a class="-x-ago" href="" title="%%neon-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                               )
                             </span>
                             <span class="packages-metadata-block">
