+++ app/test/frontend/golden/publisher_unlisted_packages_page.html
@@ -131,7 +131,7 @@
                   </p>
                   <p>
                     Publisher registered
-                    <a class="-x-ago" href="" title="on %%neon-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                    <a class="-x-ago" href="" title="on %%neon-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                   </p>
                 </div>
               </div>
@@ -234,7 +234,7 @@
                               v
                               <a href="/packages/neon">1.0.0</a>
                               (
-                              <a class="-x-ago" href="" title="%%neon-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                              <a class="-x-ago" href="" title="%%neon-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                               )
                             </span>
                             <span class="packages-metadata-block">
@@ -308,7 +308,7 @@
                               v
                               <a href="/packages/flutter_titanium">1.10.0</a>
                               (
-                              <a class="-x-ago" href="" title="%%neon-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button">%%x-ago%%</a>
+                              <a class="-x-ago" href="" title="%%neon-created-date%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%millis%%">%%x-ago%%</a>
                               )
                             </span>
                             <span class="packages-metadata-block">
