+++ app/test/task/testdata/goldens/packages/oxygen/score.html
@@ -128,7 +128,7 @@
                 <div class="metadata">
                   Published
                   <span>
-                    <a class="-x-ago" href="" title="%%short-dateformat%%" aria-label="Switch between date and elapsed time." aria-role="button">%%time-ago%%</a>
+                    <a class="-x-ago" href="" title="%%short-dateformat%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%time-ago-millis%%">%%time-ago%%</a>
                   </span>
                   <span class="package-badge" title="Package is compatible with Dart 3.">Dart 3 compatible</span>
                 </div>
@@ -227,7 +227,7 @@
                   </div>
                   <p class="analysis-info">
                     We analyzed this package
-                    <a class="-x-ago" href="" title="on %%short-dateformat%%" aria-label="Switch between date and elapsed time." aria-role="button">%%time-ago%%</a>
+                    <a class="-x-ago" href="" title="on %%short-dateformat%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%time-ago-millis%%">%%time-ago%%</a>
                     , and awarded it 110 pub points (of a possible 140):
                   </p>
                   <div class="pkg-report">
