+++ app/test/task/testdata/goldens/packages/oxygen/versions/2.0.0.html
@@ -128,7 +128,7 @@
                 <div class="metadata">
                   Published
                   <span>
-                    <a class="-x-ago" href="" title="%%short-dateformat%%" aria-label="Switch between date and elapsed time." aria-role="button">%%time-ago%%</a>
+                    <a class="-x-ago" href="" title="%%short-dateformat%%" aria-label="Switch between date and elapsed time." aria-role="button" data-timestamp="%%time-ago-millis%%">%%time-ago%%</a>
                   </span>
                   <span class="package-badge" title="Package is compatible with Dart 3.">Dart 3 compatible</span>
                 </div>
